flask-graphql-mysql/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cross-web==0.4.1
cryptography==46.0.5
ecdsa==0.19.1
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
graphql-core==3.2.8
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysqlclient==2.2.8
packaging==26.0
passlib==1.7.4
psycopg2-binary==2.9.11
pwdlib==0.3.0
pyasn1==0.6.2
pycparser==3.0
python-dateutil==2.9.0.post0
python-jose==3.5.0
pytz==2026.1.post1
rsa==4.9.1
setuptools==82.0.1
six==1.17.0
SQLAlchemy==2.0.48
sqlalchemy-serializer==1.6.2
strawberry-graphql==0.311.1
typing_extensions==4.15.0
Werkzeug==3.1.6
